Motion in modulus

+++ src/motionstage/src/mylistener.py
@@ -7,6 +7,9 @@ from std_srvs.srv import Empty,EmptyResponse
 import geometry_msgs.msg
 from geometry_msgs.msg import WrenchStamped
 import json
+import os
+import os.path
+
 # import settings
 pos_record = []
 wrench_record = []
@@ -65,16 +68,64 @@ def save_readings(req):
     filename = rospy.get_param('save_file_name')
     output_data = {'pos_list':pos_record, 'wrench_list': wrench_record }
     # save_path = '/home/mcube-daolin/catkin_ws/src/motionstage/'
-    save_path = '/home/mcube-daolin/Dropbox (MIT)/Daolin Lab Repo/Projects/2018 Engineering Friction with Micro-textures/Data Collected/'
+    date = filename.split("-")
+    year = date[0] #find the year
+    month = date[1] #find the month
+    day = date[2] #find day
+    # print("This is the year from the savepath:" + " " + year)
+    # print("This is the month from the savepath:" + " " + month)
+    # print("This is the day from the savepath:" + " " + day)
+    save_path = '/home/mcube-daolin/Dropbox (MIT)/Daolin Lab Repo/Projects/2018 Engineering Friction with Micro-textures/Data Collected/' + year + "-" + month + "-" + day + '/'
     filename = save_path + filename
-    # print("saving file")
-    # print(output_data)
-    # print(type(output_data))
-    with open(filename, 'w') as outfile:  # write data to json file
-        # print(type(outfile))
-        print(filename)
-        json.dump(output_data, outfile)   #TODO: find out why failing to save the file.
-        outfile.close()
+
+    if not os.path.isdir(save_path):#if the directory doesn't exist make one and save the file inside
+        print("Directory doesn't exist")
+        os.mkdir(save_path)
+        # print("Savepath you created: ")
+        # print(repr(save_path))
+        print("New Directory Created")
+        #maybe i can just add the next 2 lines of code after this function  is called
+        # file_num += 1
+        # rospy.set_param('save_file_number',file_num)
+        with open(filename, 'w') as outfile:  # write data to json file
+            # print(type(outfile))
+            print(filename)
+            json.dump(output_data, outfile)   #TODO: find out why failing to save the file.
+            outfile.close()
+    else: #if the path already exists write the file in the existing folder
+        print("Directory Already Exists")
+        print('old path: ' + filename)
+        #while the filename is not unique add an "a" in from of the filename
+        while os.path.isfile(filename):
+            print("File Already Exists")
+            #change the file name
+            expfile_previous = rospy.get_param('save_file_name')
+            path_split = filename.split(day + '/')
+            # print('Split path:')
+            # print(path_split)
+            filename = save_path +  'a' + path_split[1]
+            # print('new path: ' + filename)
+            # exp = 'a'+ 'save_file_name'
+        with open(filename, 'w') as outfile:
+                json.dump(output_data, outfile)
+                outfile.close()
+    #    print(number_files)
+    #    number_files += 1
+    #    filename  = filename + str(number_files)
+    #    print("Number of files: " + str(number_files))
+    # #    print("Filename: " + filename)
+    #     if not os.path.isfile(filename): #if the file doesn't exist already, write
+    #         with open(filename, 'w') as outfile:  # write data to json file
+    #         # print(type(outfile))
+    # #        print(filename)
+    #             json.dump(output_data, outfile)   #TODO: find out why failing to save the file.
+    #             outfile.close()
+    #     else: #if the file exists already give it a new name
+    #         with open(filename, 'w') as outfile:
+    #             #(append)new_output_data = str(output_data) #Q: is this format acceptable?
+    #             #(append) outfile.write(new_output_data) #outputing string of new data to old file
+    #             json.dump(output_data, outfile)
+    #             outfile.close()
 
     # print("file saved")
     rospy.sleep(.3)
@@ -85,7 +136,7 @@ if __name__ == '__main__':
     try:
         rospy.init_node('lisener_node', log_level = rospy.INFO)
         s_1 = rospy.Service('start_read', Empty, start_read)
-        s_1 = rospy.Service('save_readings', Empty, save_readings)
+        s_2 = rospy.Service('save_readings', Empty, save_readings)
         print ('mylistener ready!')
         exp_listener()
     except rospy.ROSInterruptException:
